DONE!! Debugged and dundered. I'm pretty sure it's slick. Gonna make sure that it all works past one load. Learned the ins and outs of KVO, that's for sure. I feel like I have a huge separation in my ability to learn in the classroom and the speed needs of the course. I'm also a ridiculously slow programmer, but I think a lot of that is nerves and being worn down in general. I have a much easier time on the build weeks because I get more sleep in general. I'm excited for my hiatus. It's gonna be nice to catch up on sleep. I look really rough right now.
Update: There's a bug that appears that I am not going to deal with in that a pokemon may not have a front_shiny version. I could implement the fix, but if I did that it would pretty much mean that I should just use front_default instead. It's known, I figured it out by debugging, gimme the bonus points for that. Just using the shiny versions cause it's the closest I am ever getting to a shiny charizard. I also limited the call to the standard 20. That's mot a bug, that's me running this 20000000000000x for the KVO and wanting a faster loading time.

The app also runs kind of natively smooth memory wise. It initiates the pokemon then passes it to the other controller, which is dismissed (I believe) when exiting due to it being a detail view presentation.

Going to SLEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEP